build: integrate commitlint and husky
I added commitlint and husky to the project using NPM. I will use
commitlint to validate that commit messages conform to the Conventional
Commits specification and project formatting rules. I will use husky to
automate running commitlint when the user attempts to create a commit to
ensure that the commit message is valid before the commit is completed
and recorded in the local repository.

I created the bin/setup.sh script to automate setting up a new clone of
the repository for development. setup.sh will install the tools using
NPM and will install husky with the local Git repository to execute the
hook programs.

I started creating documentation describing the software requirements
for developing new features or fixing bugs in CommonMarkSwift. I created
a second document describing how to run the setup.sh script to configure
the local repository.